+++ src/FlomtManager.Data.EF/Entities/ParameterEntity.cs
@@ -8,7 +8,7 @@ namespace FlomtManager.Data.EF.Entities
     {
         public required byte Number { get; set; }
         public required ParameterType ParameterType { get; set; }
-        public required float Comma { get; set; }
+        public required byte Comma { get; set; }
         public required ushort ErrorMask { get; set; }
         public required byte IntegrationNumber { get; set; }
         [MaxLength(4)]
